Type information in exceptions without context
A number of exceptions was thrown without any context of where the error
occurs. This makes it very hard to find the cause when (de)serializing
large object trees.

Type/member information has been added to the message in a similar
fashion as other exception messages which already included this.